Import of ISHIGAKI/Catalyst-Authentication-Store-Jifty-DBI-0.02 from CPAN.
gitpan-cpan-distribution: Catalyst-Authentication-Store-Jifty-DBI
gitpan-cpan-version:      0.02
gitpan-cpan-path:         ISHIGAKI/Catalyst-Authentication-Store-Jifty-DBI-0.02.tar.gz
gitpan-cpan-author:       ISHIGAKI
gitpan-cpan-maturity:     released

+++ t/lib/TestApp/Controller/DB.pm
@@ -74,11 +74,11 @@ sub teardown : Local {
   my ($self, $c) = @_;
 
   my $dbfile = $c->model('TestDB')->database;
-  if ( -f $dbfile ) {
+  if ( $dbfile && -f $dbfile ) {
     $c->model('TestDB')->disconnect;
     unlink $dbfile;
   }
-  $c->detach('/error') if -f $dbfile;
+  $c->detach('/error') if $dbfile && -f $dbfile;
 
   $c->res->body('ok');
 }
